+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 * Elixir v0.14 support
 * Add Travis CI
+* Added `Faker.Company`
+  * `bs/0`
+  * `bullshit/0`
+  * `bullshit_prefix/0`
+  * `bullshit_suffix/0`
+  * `buzzword/0`
+  * `buzzword_prefix/0`
+  * `buzzword_suffix/0`
+  * `catch_phrase/0`
+  * `name/0`
+  * `suffix/0`
 
 # 0.2.0